requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
future==1.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyposdisplay==0.0.8
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0.2
Unidecode==1.3.8
Werkzeug==3.1.3
